refactor: Lift modal state
